+++ src/vbComponents/elements/BulmaImage.vue
@@ -9,12 +9,20 @@
 	</figure>
 </template>
 
+<style lang="css" scoped>
+    figure.image {
+        width: v-bind(width);
+    }
+</style>
+
 <script lang="ts" setup>
 	import {computed} from "vue";
 	import {getBulmaClassesFromProps} from "../../types";
 	import {BulmaAspectRatio, BulmaDimension} from "../../types/ImageTypes";
 
 	const props = defineProps<{
+		/** css width of image container (figure) */
+		width?: string
 		/** Standard dimensions for the image container to be from 16 to 128 increase by a factor of 2^x */
 		dimensions?: BulmaDimension
 		/** If the container should be rounder; clipping whatever does not fit into the image
@@ -24,8 +32,7 @@
 		 * @default false*/
 		isRounded?: boolean
 		/** The aspect ratio to apply to the image or content.
-		 * Recommended to have the parent to have a set width or the content will expand to the full width of the parent
-		 */
+		 * Recommended to have the parent to have a set width ({@link width}) or the content will expand to the full width of the parent*/
 		aspectRatio?: BulmaAspectRatio
 		/** If the image doesn't already expand to the parents' width, set the width to expand to available space
 		 * @default false*/